fixed profile page

+++ internal/handlers/post/user-submissions.go
@@ -69,7 +69,7 @@ func GetUserSubmissions(w http.ResponseWriter, r *http.Request) {
 	for _, doc := range returnedProfile {
 		for key, value := range doc {
 			if key == "userId" {
-				userID = value.(int64)
+				userID = int64(value.(float64))
 				break
 			}
 		}
